+++ built/jtree.browser.js
@@ -678,6 +678,9 @@ class ImmutableNode extends AbstractNode {
         const content = this.getContent();
         return (content ? content : "") + (this.length ? this.getYI() + this._childrenToString() : "");
     }
+    getFirstNode() {
+        return this.nodeAt(0);
+    }
     getStack() {
         return this._getStack();
     }
@@ -880,8 +883,12 @@ class ImmutableNode extends AbstractNode {
     toHtml() {
         return this._childrenToHtml(0);
     }
+    _getHtmlJoinByCharacter() {
+        return `<span class="yIncrement">${this.getYI()}</span>`;
+    }
     _childrenToHtml(indentCount) {
-        return this.map(node => node._toHtml(indentCount)).join(`<span class="yIncrement">${this.getYI()}</span>`);
+        const joinBy = this._getHtmlJoinByCharacter();
+        return this.map(node => node._toHtml(indentCount)).join(joinBy);
     }
     _childrenToString(indentCount, language = this) {
         return this.map(node => node.toString(indentCount, language)).join(language.getYI());
@@ -958,23 +965,6 @@ class ImmutableNode extends AbstractNode {
     _childrenToYamlAssociativeArray(indentLevel) {
         return this.map(node => node._toYamlAssociativeArrayElement(indentLevel));
     }
-    // todo: do we need this?
-    _getDuplicateLinesMap() {
-        const count = {};
-        this.forEach(node => {
-            const line = node.getLine();
-            if (count[line])
-                count[line]++;
-            else
-                count[line] = 1;
-        });
-        this.forEach(node => {
-            const line = node.getLine();
-            if (count[line] === 1)
-                delete count[line];
-        });
-        return count;
-    }
     toJson() {
         return JSON.stringify(this.toObject(), null, " ");
     }
@@ -1001,14 +991,6 @@ class ImmutableNode extends AbstractNode {
             node.destroy();
         });
         return clone;
-    }
-    isLeafColumn(path) {
-        for (let node of this._getChildren()) {
-            const nd = node.getNode(path);
-            if (nd && nd.length)
-                return false;
-        }
-        return true;
     }
     getNode(firstWordPath) {
         return this._getNodeByPath(firstWordPath);
@@ -1117,14 +1099,6 @@ class ImmutableNode extends AbstractNode {
     }
     toCsv() {
         return this.toDelimited(",");
-    }
-    toFlatTree() {
-        const tree = this.clone();
-        tree.forEach(node => {
-            // todo: best approach here? set children as content?
-            node.deleteChildren();
-        });
-        return tree;
     }
     _getTypes(header) {
         const matrix = this._getMatrix(header);
@@ -1415,8 +1389,6 @@ class ImmutableNode extends AbstractNode {
         return newNode;
     }
     _parseString(str) {
-        if (!str)
-            return this;
         const lines = str.split(this.getYIRegex());
         const parentStack = [];
         let currentIndentCount = -1;
@@ -1463,6 +1435,7 @@ class ImmutableNode extends AbstractNode {
         const result = this._getIndex()[firstWord];
         return result === undefined ? -1 : result;
     }
+    // todo: renmae to indexOfFirst?
     indexOf(firstWord) {
         if (!this.has(firstWord))
             return -1;
@@ -1472,7 +1445,6 @@ class ImmutableNode extends AbstractNode {
             if (nodes[index].getFirstWord() === firstWord)
                 return index;
         }
-        return -1;
     }
     toObject() {
         return this._toObject();
@@ -1511,16 +1483,6 @@ class ImmutableNode extends AbstractNode {
     }
     _hasFirstWord(firstWord) {
         return this._getIndex()[firstWord] !== undefined;
-    }
-    // todo: is this used anywhere?
-    _getFirstWordByIndex(index) {
-        // Passing -1 gets the last item, et cetera
-        const length = this.length;
-        if (index < 0)
-            index = length + index;
-        if (index >= length)
-            return undefined;
-        return this._getChildren()[index].getFirstWord();
     }
     map(fn) {
         return this.getChildren().map(fn);